security/requirements.txt
# Shared security library dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
pyyaml>=6.0.1
httpx>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6